fixing bugs

+++ models/engine/db_storage.py
@@ -12,6 +12,12 @@ from models.place import Place
 from models.review import Review
 from models.amenity import Amenity
 
+user = os.getenv("HBNB_MYSQL_USER")
+password = os.getenv("HBNB_MYSQL_PWD")
+host = os.getenv("HBNB_MYSQL_HOST")
+database = os.getenv("HBNB_MYSQL_DB")
+env = os.getenv("HBNB_ENV")
+
 
 class DBStorage():
     """ Class to manage storage of hbnb models in MySQL database """
@@ -20,16 +26,17 @@ class DBStorage():
 
     def __init__(self):
         """ Create DBStorage instance """
-        self.__engine = create_engine('mysql+mysqldb://{}:{}@{}/{}'.format(
-            getenv("HBNB_MYSQL_USER"),
-            getenv("HBNB_MYSQL_PWD"),
-            getenv("HBNB_MYSQL_HOST"),
-            getenv("HBNB_MYSQL_DB"),
-            pool_pre_ping=True
-        ))
 
-        # if getenv("HBNB_ENV") == "test":
-              Base.metadata.drop_all(self.__engine)
+	self.__engine = create_engine(
+            f'mysql+mysqldb://{user}:{password}@{host}/{database}',
+            pool_pre_ping=True
+        )
+
+	 if env == "test":
+            Base.metadata.drop_all(self.__engine)
+
+        Session = sessionmaker(bind=self.__engine, expire_on_commit=False)
+        self.__session = scoped_session(Session)
 
     def all(self, cls=None):
         """Queries the database based on the class name"""
